Add "Array vs TypedArray Dot Product"
Add example for calculating dot products of two vectors within a large array of vector elements (as might be used in a "Struct of Arrays" implementation.) The example begins by generating an array of random numbers as vector elements, and copying that into both a native array and a Float64Array. It then uses the `pair` function to convert a number within [0, N) to the pair of indices [i, N-i-1] * 3. These pair are the reduction value between iterations. The initial seed is from Math.random() * N; later seeds are from the dot product cos(t) = A*B/|A||B|. This operation requires 10 multiplications, 6 additions, 2 square roots, and 1 division. The core intent is to evaluate the speed of these operations between native arrays and TypeArrays.